Update FingeringWavelengthtesting.py
Most of the correlation finding is now working. Deleted a bit of the old code. Still need to find a bit of the code and make a simple wrapper function so it can be run on an image sequence as well as seeing if the center location remains constant (it should).